# docs/architecture/openapi/speakeasy-api.yaml
openapi: 3.0.3
info:
  title: SpeakEasy API
  version: 0.1.0
  description: 'Machine-readable OpenAPI source of truth for SpeakEasy server-backed Product Base behavior, P0 commercial
    subscription readiness, and P0.1 expression automation training. P0.2/P1/P2 roadmap items are intentionally
    excluded from implementation-level paths until they have owning increments.

    '
  license:
    name: Proprietary
servers:
- url: /v1
  description: Versioned API base path
security:
- bearerAuth: []
...
paths:
  /auth/login/phone:
    post:
      tags:
      - Auth
      operationId: loginPhone
      summary: Login with phone verification code.
      security: []
      x-traceability:
        product_objects:
        - product-base
        - p0-commercial-readiness
        requirements:
        - FR-001
        - FR-COM-004
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhoneLoginRequest'
            examples:
              default:
                $ref: '#/components/examples/PhoneLoginRequestExample'
      responses:
        '200':
          description: Auth session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
              examples:
                default:
                  $ref: '#/components/examples/AuthSessionResponseExample'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /auth/login/apple:
    post:
      tags:
      - Auth
      operationId: loginApple
      summary: Login with Apple identity token.
      security: []
      x-traceability:
        product_objects:
        - product-base
        - p0-commercial-readiness
        requirements:
        - FR-001
        - FR-COM-005
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialLoginRequest'
            examples:
              default:
                $ref: '#/components/examples/SocialLoginRequestExample'
      responses:
        '200':
          description: Auth session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
              examples:
                default:
                  $ref: '#/components/examples/AuthSessionResponseExample'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '503':
          $ref: '#/components/responses/ProviderUnavailable'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /auth/login/wechat:
    post:
      tags:
      - Auth
      operationId: loginWechat
      summary: Login with WeChat authorization code.
      security: []
      x-traceability:
        product_objects:
        - product-base
        - p0-commercial-readiness
        requirements:
        - FR-001
        - FR-COM-005
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialLoginRequest'
            examples:
              default:
                $ref: '#/components/examples/SocialLoginRequestExample'
      responses:
        '200':
          description: Auth session created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
              examples:
                default:
                  $ref: '#/components/examples/AuthSessionResponseExample'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '503':
          $ref: '#/components/responses/ProviderUnavailable'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /auth/refresh:
    post:
      tags:
      - Auth
      operationId: refreshAuth
      summary: Refresh access token.
      security: []
      x-traceability:
        product_objects:
        - product-base
        - p0-commercial-readiness
        requirements:
        - FR-001
        - FR-COM-004
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            examples:
              default:
                $ref: '#/components/examples/RefreshTokenRequestExample'
      responses:
        '200':
          description: Token refreshed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
              examples:
                default:
                  $ref: '#/components/examples/AuthSessionResponseExample'
        '401':
          $ref: '#/components/responses/Unauthenticated'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /auth/logout:
    post:
      tags:
      - Auth
      operationId: logout
      summary: Logout current session.
      x-traceability:
        product_objects:
        - product-base
        requirements:
        - FR-001
        - FR-010
      responses:
        '204':
          description: Logout completed.
        '401':
          $ref: '#/components/responses/Unauthenticated'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /user/me:
    get:
      tags:
      - User
      operationId: getMe
      summary: Get current user profile and account state.
      x-traceability:
        product_objects:
        - product-base
        - p0-commercial-readiness
        requirements:
        - FR-010
        - FR-COM-004
      responses:
        '200':
          description: Current user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
              examples:
                default:
                  $ref: '#/components/examples/UserProfileResponseExample'
        '401':
          $ref: '#/components/responses/Unauthenticated'
    patch:
      tags:
      - User
      operationId: updateMe
      summary: Update user profile preferences.
      x-traceability:
        product_objects:
        - product-base
        requirements:
        - FR-010
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
            examples:
              default:
                $ref: '#/components/examples/UpdateUserProfileRequestExample'
      responses:
        '200':
          description: Updated profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
              examples:
                default:
                  $ref: '#/components/examples/UserProfileResponseExample'
        '422':
          $ref: '#/components/responses/ValidationFailed'
    delete:
      tags:
      - User
      operationId: requestAccountDeletion
      summary: Request account deletion and cloud data cleanup.
      x-idempotency-required: true
      x-traceability:
        product_objects:
        - product-base
        - p0-commercial-readiness
        requirements:
        - FR-010
        - FR-COM-008
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Deletion job accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDeletionJobResponse'
              examples:
                default:
                  $ref: '#/components/examples/AccountDeletionJobResponseExample'
        '409':
          $ref: '#/components/responses/Conflict'
    parameters:
    - $ref: '#/components/parameters/RequestId'
  /user/deletion-status:
    get:
      tags:
      - User
      operationId: getAccountDeletionStatus
      summary: Get latest account deletion job status.
      x-traceability:
        product_objects:
        - mvp-backend-membership-boundary
        requirements:
        - MVP-BE-FR-011
      responses:
        '200':
          description: Latest deletion job status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDeletionJobResponse'
              examples:
                default:
                  $ref: '#/components/examples/AccountDeletionJobResponseExample'
        '404':
          $ref: '#/components/responses/NotFound'
    parameters:
    - $ref: '#/components/parameters/RequestId'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    opsBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Restricted operations token.
    webhookSignature:
      type: apiKey
      in: header
      name: X-Provider-Signature
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              $ref: '#/components/examples/ErrorResponseExample'
    Unauthenticated:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              $ref: '#/components/examples/ErrorResponseExample'
...
    ValidationFailed:
      description: Schema validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              $ref: '#/components/examples/ErrorResponseExample'
    ProviderUnavailable:
      description: External provider unavailable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              $ref: '#/components/examples/ErrorResponseExample'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              $ref: '#/components/examples/ErrorResponseExample'
    Conflict:
      description: Current state conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            default:
              $ref: '#/components/examples/ErrorResponseExample'
  schemas:
    SchemaVersion:
      type: integer
      enum:
      - 1
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - request_id
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
    ErrorCode:
      type: string
      enum:
      - UNAUTHENTICATED
      - FORBIDDEN
      - ENTITLEMENT_REQUIRED
      - USAGE_LIMIT_EXCEEDED
      - INVALID_RECEIPT
      - PRODUCT_MISMATCH
      - SUBSCRIPTION_EXPIRED
      - IDEMPOTENCY_CONFLICT
      - SCHEMA_VALIDATION_FAILED
      - PROVIDER_UNAVAILABLE
      - DELETE_IN_PROGRESS
      - RESOURCE_NOT_FOUND
      - CONFLICT
    LevelCode:
      type: string
      enum:
      - L1
      - L2
      - L3
...
    AuthSessionResponse:
      type: object
      required:
      - schema_version
      - user
      - access_token
      - refresh_token
      - expires_at
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        user:
          $ref: '#/components/schemas/UserProfile'
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
    PhoneLoginRequest:
      type: object
      required:
      - schema_version
      - phone_number
      - verification_code
      - terms_accepted
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        phone_number:
          type: string
        verification_code:
          type: string
        terms_accepted:
          type: boolean
    SocialLoginRequest:
      type: object
      required:
      - schema_version
      - provider_token
      - terms_accepted
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        provider_token:
          type: string
        nonce:
          type: string
        terms_accepted:
          type: boolean
    RefreshTokenRequest:
      type: object
      required:
      - schema_version
      - refresh_token
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        refresh_token:
          type: string
    UserProfile:
      type: object
      required:
      - user_id
      - account_status
      - onboarding_status
      properties:
        user_id:
          type: string
        display_name:
          type: string
        avatar_ref:
          type: string
          nullable: true
        locale:
          type: string
          default: zh-CN
        target_level:
          $ref: '#/components/schemas/LevelCode'
        daily_minutes:
          type: integer
          minimum: 1
        account_status:
          type: string
          enum:
          - active
          - deletion_requested
          - deleted
          - disabled
        onboarding_status:
          type: string
          enum:
          - incomplete
          - complete
    UserProfileResponse:
      type: object
      required:
      - schema_version
      - user
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        user:
          $ref: '#/components/schemas/UserProfile'
    UpdateUserProfileRequest:
      type: object
      required:
      - schema_version
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        display_name:
          type: string
        target_level:
          $ref: '#/components/schemas/LevelCode'
        daily_minutes:
          type: integer
          minimum: 1
        reminder_enabled:
          type: boolean
        reminder_time:
          type: string
          pattern: ^([01][0-9]|2[0-3]):[0-5][0-9]$
    AccountDeletionJobResponse:
      type: object
      required:
      - schema_version
      - deletion_job_id
      - status
      properties:
        schema_version:
          $ref: '#/components/schemas/SchemaVersion'
        deletion_job_id:
          type: string
        status:
          type: string
          enum:
          - requested
          - access_revoked
          - deleting_learning_data
          - anonymizing_audit_refs
          - completed
          - failed
        requested_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        failure_reason:
          type: string
          nullable: true
...
  examples:
    ErrorResponseExample:
      summary: Standard error response
      value:
        error:
          code: SCHEMA_VALIDATION_FAILED
          message: Request validation failed.
          request_id: req_20260526_sample
          details:
            field: schema_version
            reason: Unsupported schema version.
    PhoneLoginRequestExample:
      summary: loginPhone request example
      value:
        schema_version: 1
        phone_number: '+8613800138000'
        verification_code: SAMPLE_CODE
        terms_accepted: true
    AuthSessionResponseExample:
      summary: loginPhone response 200 example
      value:
        schema_version: 1
        user:
          user_id: user_id_sample
          display_name: Sample Name
          avatar_ref: sample
          locale: zh-CN
          target_level: L1
          daily_minutes: 1
          account_status: active
          onboarding_status: incomplete
        access_token: sample-token
        refresh_token: sample-token
        expires_at: '2026-05-26T12:00:00Z'
    SocialLoginRequestExample:
      summary: loginApple request example
      value:
        schema_version: 1
        provider_token: sample-token
        nonce: sample
        terms_accepted: true
    RefreshTokenRequestExample:
      summary: refreshAuth request example
      value:
        schema_version: 1
        refresh_token: sample-token
    UserProfileResponseExample:
      summary: getMe response 200 example
      value:
        schema_version: 1
        user:
          user_id: user_id_sample
          display_name: Sample Name
          avatar_ref: sample
          locale: zh-CN
          target_level: L1
          daily_minutes: 1
          account_status: active
          onboarding_status: incomplete
    UpdateUserProfileRequestExample:
      summary: updateMe request example
      value:
        schema_version: 1
        display_name: Sample Name
        target_level: L1
        daily_minutes: 1
        reminder_enabled: true
        reminder_time: 09:00
    AccountDeletionJobResponseExample:
      summary: requestAccountDeletion response 202 example
      value:
        schema_version: 1
        deletion_job_id: deletion_job_id_sample
        status: completed
        requested_at: '2026-05-26T12:00:00Z'
        completed_at: '2026-05-26T12:00:00Z'
        failure_reason: null
...